added npm package setup
Mentioned the package's name on npm.

This is important as it is not easy to find reveal.js on npm: when [searching for "reveal"](https://www.npmjs.com/search?q=reveal&page=1&ranking=optimal), the first result is a package called `reveal`, but  published by someone else...  Apparently a fork of v0.0.4. `reveal.js` does not appear at all in the first two result pages.